refactor cast spell
