+++ module/actor/mixins/baseDamageMixin.js
@@ -1,0 +1,13 @@
+export let baseDamageMixin = {
+    createBaseDamageObject(item) {
+        return {
+            damageProperties: item.system.damageProperties,
+            item: item,
+            itemUuid: item.uuid,
+            crit: {
+                critLocationModifier: item.parent.system.attackStats.critLocationModifier,
+                critEffectModifier: item.parent.system.attackStats.critEffectModifier
+            }
+        };
+    }
+};
