+++ module/actor/witcherActor.js
@@ -3,6 +3,8 @@ import { getRandomInt } from '../scripts/helper.js';
 import { RollConfig } from '../scripts/rollConfig.js';
 import { WITCHER } from '../setup/config.js';
 import { modifierMixin } from './mixins/modifierMixin.js';
+import { baseDamageMixin } from './mixins/baseDamageMixin.js';
+import { castSpellMixin } from './mixins/castSpellMixin.js';
 
 export default class WitcherActor extends Actor {
     prepareDerivedData() {
@@ -398,7 +400,7 @@ export default class WitcherActor extends Actor {
         return Math.max(encumbranceModifier, 0);
     }
 
-    async rollItem(itemId) {
+    async useItem(itemId) {
         let item = this.items.get(itemId);
 
         if (!item) return;
@@ -408,7 +410,7 @@ export default class WitcherActor extends Actor {
         }
 
         if (item.type === 'spell') {
-            this.rollSpell(itemId);
+            this.castSpell(item);
         }
 
         if (item.system.isConsumable) {
@@ -419,10 +421,6 @@ export default class WitcherActor extends Actor {
 
     async rollWeapon(itemId) {
         this.sheet._onItemRoll(null, itemId);
-    }
-
-    async rollSpell(itemId) {
-        this.sheet._onSpellRoll(null, itemId);
     }
 
     async rollSkill(skillName) {
@@ -840,3 +838,5 @@ export default class WitcherActor extends Actor {
 }
 
 Object.assign(WitcherActor.prototype, modifierMixin);
+Object.assign(WitcherActor.prototype, baseDamageMixin);
+Object.assign(WitcherActor.prototype, castSpellMixin);
